Attempting to extract logic from the chain service
